# adminConsoleBackend/src/auth/dto/login.dto.ts
import { IsEmail, IsNotEmpty, IsString } from "class-validator";

export class LoginDTO{
    @IsEmail()
    @IsNotEmpty()
    email:string;

    @IsString()
    @IsNotEmpty()
    password:string;
}