# adminConsoleBackend/src/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/ping":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/devices":{"get":{"description":"Get information of the device using the searchstring, os type, device type and device status ","operationId":"DevicesController_getAllDevices","parameters":[{"name":"deviceType","required":false,"in":"query","description":"Device Type of the device in number: Others = 1, Mobile = 2","schema":{"example":1,"type":"number","enum":[1,2]}},{"name":"deviceStatus","required":false,"in":"query","description":"Device Status of the device in number: Active = 1, dead/unassigned = 2","schema":{"example":1,"type":"number","enum":[1,2]}},{"name":"page","required":true,"in":"query","description":"Page Number To be shown","schema":{"example":1,"type":"number"}},{"name":"limit","required":true,"in":"query","description":"Limit in per page","schema":{"example":1,"type":"number"}},{"name":"osType","required":false,"in":"query","description":"Os Type of the device in number: Linux = 1, MAC = 2, Windows = 3, Android = 4, IOS = 5","schema":{"example":3,"type":"number","enum":[1,2,3,4,5]}},{"name":"searchString","required":false,"in":"query","description":"Search string too serach devices using the serial number...","schema":{"example":"NIUF794WTH","type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Get information of the device","tags":["Devices"]},"post":{"description":"Insert the information abount the  new device that is not present before ","operationId":"DevicesController_createDevice","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}}}},"security":[{"JWT-auth":[]}],"summary":"Insert new device information","tags":["Devices"]}},"/api/devices/status":{"patch":{"description":"Updating ythe device status to dead or active state. If dead state removing the user Id","operationId":"DevicesController_updateDeviceStatus","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDeviceDTO"}}}},"responses":{"202":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"JWT-auth":[]}],"summary":"Updating the device status","tags":["Devices"]}},"/api/devices/count":{"get":{"description":"Count the number of device by device type, device status andos type.","operationId":"DevicesController_countByStatusAndDeviceType","parameters":[{"name":"deviceStatus","required":true,"in":"query","description":"Status of the device","schema":{"enum":[1,2],"type":"number"}},{"name":"deviceType","required":true,"in":"query","description":"Type of the device","schema":{"enum":[1,2],"type":"number"}},{"name":"osType","required":false,"in":"query","description":"Type of the os in the device","schema":{"enum":[1,2,3,4,5],"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Count the total number of device","tags":["Devices"]}},"/api/devices/getInventory":{"get":{"description":"Get information of the active device and the person to whome it is assigned","operationId":"DevicesController_getInventory","parameters":[{"name":"page","required":true,"in":"query","schema":{"example":1,"type":"number"}},{"name":"limit","required":true,"in":"query","description":"Total data per page","schema":{"example":1,"type":"number"}},{"name":"searchString","required":false,"in":"query","description":"Search value like serial number, name and email","schema":{"example":"6Q","type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Get information of the active device and the person to whome it is assigned","tags":["Devices"]}},"/api/devices/{deviceId}":{"delete":{"operationId":"DevicesController_deleteDevice","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Devices"]}},"/api/devices/editDevice":{"put":{"operationId":"DevicesController_editDevice","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}}},"responses":{"200":{"description":""}},"tags":["Devices"]}},"/api/devices/addBitlocker":{"put":{"operationId":"DevicesController_addBitLockerKey","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BitLockerAndRecovaryKey"}}}},"responses":{"200":{"description":""}},"tags":["Devices"]}},"/api/users/create":{"post":{"description":"Create a new employee in the database.","operationId":"UsersController_registerUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}},"security":[{"JWT-auth":[]}],"summary":"Create new employees","tags":["Users"]}},"/api/users/employee/active":{"get":{"description":"Get the information of all the employees and search for a particular employee.","operationId":"UsersController_allEmployees","parameters":[{"name":"userType","required":true,"in":"query","description":"Role of the user","schema":{"example":[1,2],"type":"array","items":{"type":"number","enum":[1,2,3]}}},{"name":"limit","required":true,"in":"query","description":"limit of per page","schema":{"type":"number"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"searchString","required":false,"in":"query","description":"String to be searched based on the basis of name, email and employee id","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Get all the employees","tags":["Users"]}},"/api/users/employee/edit":{"put":{"operationId":"UsersController_updateEmployee","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}},"security":[{"JWT-auth":[]}],"tags":["Users"]}},"/api/users/{userId}":{"delete":{"description":"Delete the employee and release all devices assigned to him","operationId":"UsersController_deleteEmployee","parameters":[{"name":"userId","required":true,"in":"path","description":"Id of the employee","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Delete the employee","tags":["Users"]}},"/api/users/userInfo":{"get":{"description":"Fetch the employee details by email id","operationId":"UsersController_fetchEmployeeById","parameters":[{"name":"email","required":true,"in":"query","description":"Email Id of the employee","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Get the employee information","tags":["Users"]}},"/api/users/changeTheRoleToAdmin":{"patch":{"operationId":"UsersController_updateToAdmin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDTO"}}}},"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"tags":["Users"]}},"/api/users/admin/revoke":{"patch":{"operationId":"UsersController_revokeAdminPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDTO"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/v2/users/create":{"post":{"operationId":"UsersController_registerAppUser_v2","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/api/UserDevice/inventory":{"get":{"operationId":"UsersDevicesController_getInventory","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"searchString","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UsersDevices"]}},"/api/UserDevice/assignDevice":{"post":{"operationId":"UsersDevicesController_assingDevice","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserToDevice"}}}},"responses":{"201":{"description":""}},"tags":["UsersDevices"]}},"/api/UserDevice/deleteAsssets/{userId}":{"delete":{"operationId":"UsersDevicesController_deleteAssignedDevicesOfUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UsersDevices"]}},"/api/UserDevice/removeUser/{assetId}":{"patch":{"operationId":"UsersDevicesController_releaseUser","parameters":[{"name":"assetId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UsersDevices"]}},"/api/UserDevice/bitlockerInfo":{"get":{"operationId":"UsersDevicesController_getInfoOfDevicesAndUser","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"searchString","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UsersDevices"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/refreshLogInToken":{"post":{"operationId":"AuthController_refreshLoginToken","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/pdf/pdf/{id}":{"get":{"description":"Download the pdf with users information","operationId":"PdfContoller_streamPdf","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the device","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Downloda Pdf","tags":["pdf"]}}},"info":{"title":"Admin Cosole API Documentation","description":"These APIs are used for Network Team Admin Console Application","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"DeviceDTO":{"type":"object","properties":{"id":{"type":"string","description":"Id of the device","readOnly":true},"deviceType":{"type":"number","enum":[1,2],"description":"Device Type of the device in number: Others = 1, Mobile = 2","example":1},"deviceStatus":{"type":"number","enum":[1,2],"description":"Device Status of the device in number: Active = 1, dead/unassigned = 2","example":1},"userId":{"type":"string","description":"User Id of the device","example":"fdf44da2-68ee-4419-a874-351f290fcccf"},"macAddress":{"description":"MAC address of the device","example":["AU-90-BU-90-R8-A8"],"type":"array","items":{"type":"string"}},"hostName":{"type":"string","description":"Host Name of the device","example":"HTI879RDB "},"serialNumber":{"type":"string","description":"Serial Number of the device","example":"NIUF794WTH"},"brand":{"type":"number","enum":[1,2,3,4,5],"description":"Brand Name of the device using number: HP = 1,Apple = 2, Dell = 3, Lenovo = 4","example":1}},"required":["deviceType","macAddress","hostName","serialNumber","brand"]},"UpdateDeviceDTO":{"type":"object","properties":{"deviceStatus":{"type":"number","enum":[1,2],"description":"Device Status of the device in number: Active = 1, dead/unassigned = 2","example":1},"id":{"type":"string","description":"Id of the device","example":"fdf44da2-68ee-4419-a874-351f290fcccf"}},"required":["id"]},"BitLockerAndRecovaryKey":{"type":"object","properties":{"id":{"type":"string","description":"Id of the device","readOnly":true}}},"CreateUserDTO":{"type":"object","properties":{"employeeId":{"type":"string","description":"Employee Id of the employee","example":"ARC-5878-678"},"firstName":{"type":"string","description":"First Name of the employee","example":"Surodoy"},"middleName":{"type":"string","description":"Middle name of the employee","example":"Kumar"},"lastName":{"type":"string","description":"Last name of the employee","example":"Maity"},"email":{"type":"string","description":"Email of the employee","example":"surodoyMaity@gmail.com"},"password":{"type":"string","description":"Password of the employee"}},"required":["employeeId","firstName","email"]},"UserDTO":{"type":"object","properties":{"employeeId":{"type":"string","description":"Employee Id of the employee","example":"ARC-5878-678"},"firstName":{"type":"string","description":"First Name of the employee","example":"Surodoy"},"middleName":{"type":"string","description":"Middle name of the employee","example":"Kumar"},"lastName":{"type":"string","description":"Last name of the employee","example":"Maity"},"email":{"type":"string","description":"Email of the employee","example":"surodoyMaity@gmail.com"},"userId":{"type":"string","description":"Email of the user"}},"required":["employeeId","firstName","email","userId"]},"UpdateUserDTO":{"type":"object","properties":{"userId":{"type":"string","description":"id of the employee by system","example":"2c2993d1-58d7-4650-a429-33fcd9cd8ad3"},"employeeId":{"type":"string","description":"Employee Id of the employee","example":"ARC-5878-678"},"firstName":{"type":"string","description":"First Name of the employee","example":"Surodoy"},"middleName":{"type":"string","description":"Middle name of the employee","example":"Kumar"},"lastName":{"type":"string","description":"Last name of the employee","example":"Maity"},"email":{"type":"string","description":"Email of the employee","example":"surodoyMaity@gmail.com"},"password":{"type":"string","description":"Password the employee"}},"required":["userId"]},"UserToDevice":{"type":"object","properties":{}},"LoginDTO":{"type":"object","properties":{}}}}}